Make Float±Int explicit with .toFloat() (silence Copilot FP)
bubbleHeight is an Int (mutableIntStateOf). `Float - Int` / `Float + Int`
already compile in Kotlin (stdlib Float.minus(Int)/Float.plus(Int) return
Float) — :app:compileDebugKotlin was green throughout. This adds an explicit
.toFloat() at the two sites purely for clarity, which also silences a Copilot
Kotlin false positive so the review loop can close.

Verified in the foreground: :app:compileDebugKotlin BUILD SUCCESSFUL and
:app:testDebugUnitTest --tests *CoachMarkTourTest* => 9 tests, 0 failures.

+++ apps/android/app/src/main/java/com/kungsbackacarcommunity/app/coachmark/CoachMarkOverlay.kt
@@ -199,7 +199,7 @@ private fun CoachMarkOverlay(
             val bubbleY = if (below) {
                 localTarget.bottom + gapPx + tailHeightPx
             } else {
-                localTarget.top - gapPx - tailHeightPx - bubbleHeight
+                localTarget.top - gapPx - tailHeightPx - bubbleHeight.toFloat()
             }
 
             // Only the ABOVE placement needs the measured height; below is stable at
@@ -215,7 +215,7 @@ private fun CoachMarkOverlay(
                 bubbleWidth = bubbleWidthPx,
                 insetPerSide = spotlightRadiusPx + tailHalfPx,
             )
-            val tailY = if (below) bubbleY - tailHeightPx else bubbleY + bubbleHeight
+            val tailY = if (below) bubbleY - tailHeightPx else bubbleY + bubbleHeight.toFloat()
 
             Box(
                 modifier = Modifier
